+++ Changelog.md
@@ -8,3 +8,9 @@
 - Code BC40000
 - 'Remember Location'
 - Starting Tab
+
+05/13/2017
+• Added ListBox
+• Replaces original objects to new
+- NEW -
+• Reduced 'If' expressions in codes